server/requirements.txt
Flask==2.3.1
numpy==1.24.2
Pillow==9.4.0
requests==2.31.0
tensorflow==2.16.1
